fix : - format json return of AccountController and Contrainst delete - format json return of ScreenController and Contrainst delete - edit Movie Seeder duration - edit Screen Seeder - edit Seat Seeder - edit count in Movie Seeder - add API Resource in routes

+++ app/Http/Controllers/Api/ScreenController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Screen;
+use App\Models\Showtime;
 use Illuminate\Http\Request;
 
 class ScreenController extends Controller
@@ -19,15 +20,19 @@ class ScreenController extends Controller
             $screen->where('format', 'LIKE', '%' . $request->format . '%');
         }
 
+        if ($request->has('name')) {
+            $screen->where('name', 'LIKE', '%' . $request->name . '%');
+        }
+
         if ($request->has('is_active')) {
             $screen->where('is_active', $request->is_active);
         }
 
         return response()->json([
-                'success' => true,
-                'message' => 'Screen has been show Successfully',
-                'data' => $screen->get(),
-            ], 200);
+            'success' => true,
+            'message' => 'Screen has been show Successfully',
+            'data' => $screen->get(),
+        ], 200);
     }
 
     /**
@@ -147,11 +152,28 @@ class ScreenController extends Controller
             ], 404);
         }
 
-        $screen->delete();
-        return response()->json([
-            'success' => true,
-            'message' => "Screen has been deleted successfully",
-            'data' => $screen,
-        ], 200);
+        $hasShowtimes = Showtime::where("screen_id", $id)->exists();
+
+        if ($hasShowtimes) {
+            
+            $screen->is_active = false;
+            $screen->save();
+
+            return response()->json([
+                'success' => true,
+                'message' => "Screen has been deactive successfully",
+                'data' => $screen,
+            ], 200);
+        } 
+        else {
+
+            $screen->delete();
+
+            return response()->json([
+                'success' => true,
+                'message' => "Screen has been deleted successfully",
+                'data' => $screen,
+            ], 200);
+        }
     }
 }
